update service and repository

+++ api/products/services/products_services.py
@@ -1,6 +1,9 @@
+from datetime import datetime
+from itertools import product
+
 from motor.motor_asyncio import AsyncIOMotorDatabase
 
-from api.users.schemas.inputs import ProductInput, PatchProductInput
+from api.products.schemas.inputs import ProductInput, PatchProductInput
 from models.products import ProductsModel
 from repositories.products import ProductsRepository
 
@@ -12,6 +15,9 @@ class ProductsService:
 
     async def create_product(self, product_input: ProductInput):
         print('Creating product in service')
+        product_found = await self.products_repository.get_product_by_code(product_input.product_code)
+        if product_found:
+            raise ValueError('There is already a product with that product_code')
         product_created = await self.products_repository.create_product(product_input.model_dump())
         if product_created is None:
             raise ValueError("Product not created")
@@ -37,11 +43,16 @@ class ProductsService:
     async def get_all_products(self) -> list[ProductsModel]:
         print('Getting all products in service')
         all_products = await self.products_repository.get_all_products()
+        if all_products is None:
+            raise ValueError("There are no products")
         print(f'All products in service: {all_products}')
         return all_products
 
     async def update_product(self, product_code: int, update_data: PatchProductInput) -> ProductsModel:
         print('Updating some product data by code in service')
+        product_by_update = await self.products_repository.get_product_by_code(product_code)
+        if product_by_update is None:
+            raise ValueError("Product not found")
         product_updated = await self.products_repository.update_product(product_code, update_data.model_dump())
         if product_updated is None:
             raise ValueError("Product not updated")
@@ -50,9 +61,10 @@ class ProductsService:
 
     async def update_all_product(self, product_code: int, product_data: ProductInput) -> ProductsModel:
         print('Updating all product by code in service')
-        product = await self.get_product_by_code(product_code)
+        product = await self.products_repository.get_product_by_code(product_code)
         if product is None:
             raise ValueError("Product not found")
+        product.update_at = datetime.utcnow()
         product.product_code = product_data.product_code
         product.product_name = product_data.product_name
         product.product_category = product_data.product_category
@@ -82,6 +94,9 @@ class ProductsService:
 
     async def delete_product(self, product_code: int):
         print('Deleting product by code in service')
+        product_by_delete = await self.products_repository.get_product_by_code(product_code)
+        if product_by_delete is None:
+            raise ValueError("Product not found")
         product_deleted = await self.products_repository.delete_product(product_code)
         if product_deleted is None:
             raise ValueError("Product not deleted")
